Fixing a bug in smart cast, it should try to cast to an int first, but then if its a float it should try that.  otehrwise all ints cast to float.
Former-commit-id: 648717766232fcfaf3ac49c0e31195d33602b9e5

+++ invest_natcap/raster_utils.py
@@ -2263,7 +2263,7 @@ def get_lookup_from_csv(csv_table_uri, key_field):
 
     def smart_cast(value):
         """Attempts to cat value to a float, int, or leave it as string"""
-        cast_functions = [float, int]
+        cast_functions = [int, float]
         for fn in cast_functions:
             try:
                 return fn(value)
